Update Battleship and Circles Combined
corrected code. Before after the confirm button was clicked the grid no longer showed up. Also the circle code before allowed for overlaps. This has now been fixed and the user will only be able to place 121 circles in total because that is the max amount of grid squares. This update also got rid of redundant code. It makes sure that the circle is placed inside of the grid square before it could just be placed anywhere and that was a problem.